refactor: update search state types and reorganize imports
