+++ frontend/src/components/TabContainer.tsx
@@ -1,13 +1,12 @@
 import { useRef, useEffect, useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { useSearchState } from "@/search/context";
 
 import { Song } from "../types/song";
 import type { StoredChordSheet } from "@/storage/types";
 import ChordSheetList from "./chord-sheet-list";
 import SongViewer from "./SongViewer";
-import { SearchTab } from "@/search";
+import { useSearchState, SearchTab } from "@/search";
 import UploadTab from "./tabs/UploadTab";
 import { scrollToElement } from "../utils/scroll-utils";
 import { deleteChordSheet } from "@/storage/stores/chord-sheets/operations";
@@ -37,19 +36,19 @@ const TabContainer = ({
   const location = useLocation();
   const { searchState } = useSearchState();
   const chordDisplayRef = useRef<HTMLDivElement>(null);
-  
+
   // Local state to track the last search URL for tab switching
   const [lastSearchUrl, setLastSearchUrl] = useState<string | null>(null);
 
   // Track when we're on a search-related page and store the URL
   useEffect(() => {
     const currentPath = location.pathname + location.search;
-    
+
     // Store URL if it's a search page or an artist page (not basic app tabs)
-    if (location.pathname === '/search' || 
-        (!location.pathname.startsWith('/my-chord-sheets') && 
-         !location.pathname.startsWith('/upload') && 
-         location.pathname !== '/')) {
+    if (location.pathname === '/search' ||
+      (!location.pathname.startsWith('/my-chord-sheets') &&
+        !location.pathname.startsWith('/upload') &&
+        location.pathname !== '/')) {
       setLastSearchUrl(currentPath);
     }
   }, [location.pathname, location.search]);
@@ -60,25 +59,25 @@ const TabContainer = ({
       scrollToElement(chordDisplayRef.current);
     }
   }, [selectedSong]);
-  
+
   const navigateToSearch = () => {
     // First priority: use the stored search URL from local state
     if (lastSearchUrl) {
       navigate(lastSearchUrl);
       return;
     }
-    
+
     // Fallback: construct from current search state
     if (searchState.artist || searchState.song) {
       const params = new URLSearchParams();
       if (searchState.artist) params.set('artist', toSlug(searchState.artist));
       if (searchState.song) params.set('song', toSlug(searchState.song));
-      
+
       const searchUrl = `/search?${params.toString()}`;
       navigate(searchUrl);
       return;
     }
-    
+
     // Final fallback: go to basic search page
     navigate("/search");
   };
@@ -86,7 +85,7 @@ const TabContainer = ({
   const handleTabChange = (value: string) => {
     setActiveTab(value);
     setSelectedSong(null);
-    
+
     if (value === "upload") {
       navigate("/upload");
     } else if (value === "search") {
@@ -95,15 +94,15 @@ const TabContainer = ({
       navigate("/my-chord-sheets");
     }
   };
-  
+
   const handleSongSelect = (storedChordSheet: StoredChordSheet) => {
-    
+
     // For My Chord Sheets: Navigate directly to chord sheet page
     if (storedChordSheet.artist && storedChordSheet.title) {
       // Create URL-friendly slugs using Unicode-aware function
       const artistSlug = toSlug(storedChordSheet.artist);
       const songSlug = toSlug(storedChordSheet.title);
-      
+
       const targetUrl = `/${artistSlug}/${songSlug}`;
       // Pass a minimal Song object for navigation state
       navigate(targetUrl, {
@@ -128,7 +127,7 @@ const TabContainer = ({
       });
     }
   };
-  
+
   const handleSaveUploadedChordSheet = (content: string, title: string) => {
     // NOTE: Save functionality will be implemented with IndexedDB
     console.warn('Save functionality not yet implemented');
@@ -142,7 +141,7 @@ const TabContainer = ({
   const handleChordSheetDelete = async (songPath: string) => {
     // Find the song for user feedback
     const songToDelete = myChordSheets.find(song => song.path === songPath);
-    
+
     if (!songToDelete) {
       console.error('Song not found for deletion:', songPath);
       return;
@@ -184,84 +183,84 @@ const TabContainer = ({
 
   return (
     <Tabs value={activeTab} onValueChange={handleTabChange}>
-      <TabsList 
-        className="grid w-full grid-cols-[repeat(auto-fit,_minmax(0,_1fr))]" 
+      <TabsList
+        className="grid w-full grid-cols-[repeat(auto-fit,_minmax(0,_1fr))]"
         role="tablist"
         {...cyAttr("tabs-list")}
       >
-        <TabsTrigger 
-          value="my-chord-sheets" 
-          className="text-xs sm:text-sm" 
+        <TabsTrigger
+          value="my-chord-sheets"
+          className="text-xs sm:text-sm"
           onKeyDown={(e) => handleKeyDown(e, "my-chord-sheets")}
           {...cyAttr("tab-my-chord-sheets")}
         >
           My Chord Sheets
         </TabsTrigger>
-        <TabsTrigger 
-          value="search" 
-          className="text-xs sm:text-sm" 
+        <TabsTrigger
+          value="search"
+          className="text-xs sm:text-sm"
           onKeyDown={(e) => handleKeyDown(e, "search")}
           {...cyAttr("tab-search")}
         >
           Search
         </TabsTrigger>
-        <TabsTrigger 
-          value="upload" 
-          className="text-xs sm:text-sm" 
+        <TabsTrigger
+          value="upload"
+          className="text-xs sm:text-sm"
           onKeyDown={(e) => handleKeyDown(e, "upload")}
           {...cyAttr("tab-upload")}
         >
           Upload
         </TabsTrigger>
       </TabsList>
-      
+
       <div className="mt-4 sm:mt-6">
-      {/* Always render all tab contents, hide inactive with CSS for persistence */}
-      <div style={{ display: activeTab === "search" ? "block" : "none" }}>
-        <SearchTab
-          setMySongs={setMySongs}
-          setActiveTab={setActiveTab}
-          setSelectedSong={setSelectedSong}
-        />
-      </div>
-      <div style={{ display: activeTab === "upload" ? "block" : "none" }}>
-        <UploadTab
-          chordDisplayRef={chordDisplayRef}
-          onSaveUploadedSong={handleSaveUploadedChordSheet}
-        />
-      </div>
-      <div style={{ display: activeTab === "my-chord-sheets" ? "block" : "none" }}>
-        {selectedSong ? (
-          <SongViewer
-            song={{
-              song: selectedSong,
-              chordSheet: {
-                // NOTE: Chord sheet content loading will be implemented with IndexedDB
-                title: selectedSong.title,
-                artist: selectedSong.artist,
-                songChords: '',
-                songKey: '',
-                guitarTuning: GUITAR_TUNINGS.STANDARD,
-                guitarCapo: 0
-              }
-            }}
+        {/* Always render all tab contents, hide inactive with CSS for persistence */}
+        <div style={{ display: activeTab === "search" ? "block" : "none" }}>
+          <SearchTab
+            setMySongs={setMySongs}
+            setActiveTab={setActiveTab}
+            setSelectedSong={setSelectedSong}
+          />
+        </div>
+        <div style={{ display: activeTab === "upload" ? "block" : "none" }}>
+          <UploadTab
             chordDisplayRef={chordDisplayRef}
-            onBack={() => setSelectedSong(null)}
-            onDelete={handleChordSheetDelete}
-            onUpdate={handleChordSheetUpdate}
-            hideDeleteButton={false}
-            hideSaveButton={true}
-            isFromMyChordSheets={true}
+            onSaveUploadedSong={handleSaveUploadedChordSheet}
           />
-        ) : (
-          <ChordSheetList
-            chordSheets={myChordSheets}
-            onChordSheetSelect={handleSongSelect}
-            onDeleteChordSheet={handleChordSheetDelete}
-            onUploadClick={() => handleTabChange("upload")}
-          />
-        )}
-      </div>
+        </div>
+        <div style={{ display: activeTab === "my-chord-sheets" ? "block" : "none" }}>
+          {selectedSong ? (
+            <SongViewer
+              song={{
+                song: selectedSong,
+                chordSheet: {
+                  // NOTE: Chord sheet content loading will be implemented with IndexedDB
+                  title: selectedSong.title,
+                  artist: selectedSong.artist,
+                  songChords: '',
+                  songKey: '',
+                  guitarTuning: GUITAR_TUNINGS.STANDARD,
+                  guitarCapo: 0
+                }
+              }}
+              chordDisplayRef={chordDisplayRef}
+              onBack={() => setSelectedSong(null)}
+              onDelete={handleChordSheetDelete}
+              onUpdate={handleChordSheetUpdate}
+              hideDeleteButton={false}
+              hideSaveButton={true}
+              isFromMyChordSheets={true}
+            />
+          ) : (
+            <ChordSheetList
+              chordSheets={myChordSheets}
+              onChordSheetSelect={handleSongSelect}
+              onDeleteChordSheet={handleChordSheetDelete}
+              onUploadClick={() => handleTabChange("upload")}
+            />
+          )}
+        </div>
       </div>
     </Tabs>
   );
