--- frontend/src/search/types/searchState.ts
@@ -1,4 +1,0 @@
-/**
- * Search operation state for UI feedback
- */
-export type SearchState = 'idle' | 'loading' | 'success' | 'error';
