+++ frontend/src/search/types/searchStatus.ts
@@ -1,0 +1,4 @@
+/**
+ * Search operation status for UI feedback
+ */
+export type SearchStatus = 'idle' | 'loading' | 'success' | 'error';
